Reviewed-on: http://10.245.150.78:3000/smart-manufacture/Smart-Manufacture/pulls/66

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/vo/IcesExceptionRecordVo.java
@@ -105,7 +105,7 @@ public class IcesExceptionRecordVo {
      * 影响因子
      */
     @ExcelProperty(value = "影响因子")
-    private Long exrImpactFactor;
+    private Float exrImpactFactor;
 
     /**
      * 影响等级
